Organização da rota de Metodos de Pagamento e estrutura de pastas.
