+++ package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start",
     "server": "json-server --watch db.json --port 3001 --delay 1500",
-    "lint": "eslint src/**/*.{js,jsx,ts,tsx,json}",
+    "lint": "eslint src/**/*.{js,jsx,ts,tsx,json} && tsc --noEmit",
     "lint:fix": "eslint --fix src/**/*.{js,jsx,ts,tsx,json}",
     "format": "prettier --write src/**/*.{js,jsx,ts,tsx,css,md,json} --config ./.prettierrc"
   },
