+++ src/components/pages/records/payment_methods/Table/PopoverSubMenu/index.tsx
@@ -28,7 +28,7 @@ const ModalDetailsPaymentMethods = dynamic(() => import('../ModalDetailsPaymentM
 })
 
 // Contexts Imports
-import { ContextDrawer } from '../../../../../../contexts/contextDrawer'
+import { PaymentMethodsPageContext } from '../../../../../../contexts/pages/records'
 
 // API Services
 import { patchStatusUniquePaymentMethod } from '../../../../../../services/api'
@@ -41,7 +41,8 @@ import { IPopoverSubMenu } from './types'
 import { PaymentMethodType } from '../../../../../../@types/PaymentMethodType'
 
 const PopoverSubMenuComponent: React.FC<IPopoverSubMenu> = ({ paymentMethodID, status }) => {
-  const { handleDrawerEditPaymentMethod } = useContext(ContextDrawer)
+  const { selectPaymentMethodIdForEdit, toggleIsEditing, toggleIsLoading, disclosure } =
+    useContext(PaymentMethodsPageContext)
 
   const paymentMethodSelected = useRef<PaymentMethodType | undefined>(undefined)
 
@@ -86,6 +87,13 @@ const PopoverSubMenuComponent: React.FC<IPopoverSubMenu> = ({ paymentMethodID, s
     onOpen()
   }
 
+  const handleEditPaymentMethod = (id: string) => {
+    selectPaymentMethodIdForEdit(id)
+    toggleIsEditing()
+    toggleIsLoading()
+    disclosure.onOpen()
+  }
+
   return (
     <Fragment>
       <ModalDetailsPaymentMethods
@@ -124,7 +132,7 @@ const PopoverSubMenuComponent: React.FC<IPopoverSubMenu> = ({ paymentMethodID, s
                   colorScheme='green'
                   disabled={isLoading}
                   onClick={() => {
-                    handleDrawerEditPaymentMethod(paymentMethodID)
+                    handleEditPaymentMethod(paymentMethodID)
                   }}
                 />
               </Tooltip>
